refactor(ts): extract server api
